+++ test/httpRequestAction.spec.js
@@ -4,18 +4,52 @@ const { JsonataTransform } = require('@elastic.io/component-commons-library');
 const sinon = require('sinon');
 const { expect } = require('chai');
 const nock = require('nock');
+const fs = require('fs');
 const { messages } = require('elasticio-node');
 const logger = require('@elastic.io/component-logger')();
+const client = require('elasticio-rest-node')();
 
 const { stub } = sinon;
 
 const processAction = require('../lib/actions/httpRequestAction').process;
 
+function setSecretStub(secret) {
+  nock(process.env.ELASTICIO_API_URI)
+    .get(`/v2/workspaces/${process.env.ELASTICIO_WORKSPACE_ID}/secrets/${secret.id}`)
+    .basicAuth({
+      user: process.env.ELASTICIO_API_USERNAME,
+      pass: process.env.ELASTICIO_API_KEY,
+    })
+    .reply(200, {
+      data: {
+        id: secret.id,
+        type: 'secret',
+        attributes: secret,
+      },
+    });
+}
+
+function setNoAuthSecretStub(id) {
+  setSecretStub({ type: 'noauth', id });
+}
+
 describe('httpRequest action', () => {
   let emitter;
   let currentlyEmitting = false;
+  let originalEnv;
   beforeEach(() => {
+    originalEnv = process.env;
+    process.env = {
+
+      ...process.env,
+      ELASTICIO_API_URI: 'http://test.api.e.io',
+      ELASTICIO_WORKSPACE_ID: '12344321',
+      ELASTICIO_API_USERNAME: 'user',
+      ELASTICIO_API_KEY: 'password',
+    };
+
     sinon.restore();
+    currentlyEmitting = false;
     emitter = {
       emit: stub().returns(new Promise((resolve) => {
         // eslint-disable-next-line no-unused-expressions
@@ -29,250 +63,12 @@ describe('httpRequest action', () => {
       logger,
     };
   });
+  afterEach(() => {
+    process.env = originalEnv;
+  });
 
-  describe('oauth2 credentials', () => {
-    const msg = {
-      body: {
-        url: 'http://example.com',
-      },
-    };
-    const cfg = {
-      reader: {
-        url: 'url',
-        method: 'POST',
-        headers: [],
-      },
-      auth: {},
-    };
-    it('should fail if auth.oauth2.keys is missing', async () => {
-      cfg.auth = {
-        type: 'OAuth2',
-        oauth2: {},
-      };
-      try {
-        await processAction.call(emitter, msg, cfg);
-        throw new Error('This line should never be called because await above should throw an error');
-      } catch (err) {
-        expect(err.message).equal('cfg.auth.oauth2.keys can not be empty');
-      }
-    });
-
-    it('should fail if auth.oauth2.keys.refresh_token is missing', async () => {
-      cfg.auth = {
-        type: 'OAuth2',
-        oauth2: {
-          keys: {
-            access_token: 'token',
-          },
-        },
-      };
-      try {
-        await processAction.call(emitter, msg, cfg);
-        throw new Error('This line should never be called because await above should throw an error');
-      } catch (err) {
-        expect(err.message).equal('No refresh tokens were returned by the OAuth2 provider. Try to add access_type:offline as an additional parameter');
-      }
-    });
-
-    it('should send request with oauth2 headers, with refreshed token', async () => {
-      const refreshedToken = 'refreshed_token';
-      const tokenUri = 'http://example.com/oauth/token/';
-
-      cfg.auth = {
-        type: 'OAuth2',
-        oauth2: {
-          clientId: 'e6b02a7d-eb7e-4090-b112-f78f68cd6022',
-          clientSecret: 'e6b02a7d-eb7e-4090-b112-f78f68cd6022',
-          authUri: 'http://example.com/oauth/auth',
-          tokenUri,
-          keys: {
-            access_token: 'token',
-            token_type: 'Bearer',
-            refresh_token: 'refresh_token',
-            expires_in: 28800,
-          },
-        },
-      };
-
-      const responseMessage = {
-        access_token: refreshedToken,
-        token_type: 'Bearer',
-        refresh_token: 'refresh_token',
-        expires_in: 28800,
-      };
-
-      const refreshTokenNock = nock(tokenUri)
-        .post('/', {
-          refresh_token: cfg.auth.oauth2.keys.refresh_token,
-          grant_type: 'refresh_token',
-          client_id: cfg.auth.oauth2.clientId,
-          client_secret: cfg.auth.oauth2.clientSecret,
-        })
-        .reply((uri, requestBody) => [
-          200,
-          responseMessage,
-        ]);
-
-
-      const requestNock = nock(msg.body.url, {
-        reqheaders: {
-          Authorization: `Bearer ${refreshedToken}`,
-        },
-      })
-        .intercept('/', 'POST')
-        .reply((uri, requestBody) => [
-          200,
-          { success: true },
-        ]);
-
-      await processAction.call(emitter, msg, cfg);
-
-      expect(refreshTokenNock.isDone());
-      expect(requestNock.isDone());
-    });
-
-    it('should refresh token without `expires_in` parameter', async () => {
-      const refreshedToken = 'refreshed_token';
-      const tokenUri = 'http://example.com/oauth/token/';
-
-      cfg.auth = {
-        type: 'OAuth2',
-        oauth2: {
-          clientId: 'e6b02a7d-eb7e-4090-b112-f78f68cd6022',
-          clientSecret: 'e6b02a7d-eb7e-4090-b112-f78f68cd6022',
-          authUri: 'http://example.com/oauth/auth',
-          tokenUri,
-          keys: {
-            access_token: 'token',
-            token_type: 'Bearer',
-            refresh_token: 'refresh_token',
-          },
-        },
-      };
-
-      const responseMessage = {
-        access_token: refreshedToken,
-        token_type: 'Bearer',
-        refresh_token: 'refresh_token',
-      };
-
-      const refreshTokenNock = nock(tokenUri)
-        .post('/', {
-          refresh_token: cfg.auth.oauth2.keys.refresh_token,
-          grant_type: 'refresh_token',
-          client_id: cfg.auth.oauth2.clientId,
-          client_secret: cfg.auth.oauth2.clientSecret,
-        })
-        .reply((uri, requestBody) => [
-          200,
-          responseMessage,
-        ]);
-
-
-      const requestNock = nock(msg.body.url, {
-        reqheaders: {
-          Authorization: `Bearer ${refreshedToken}`,
-        },
-      })
-        .intercept('/', 'POST')
-        .reply((uri, requestBody) => [
-          200,
-          { success: true },
-        ]);
-
-      await processAction.call(emitter, msg, cfg);
-
-      expect(refreshTokenNock.isDone());
-      expect(requestNock.isDone());
-    });
-
-    it('should send request with oauth2 headers, without refreshed token', async () => {
-      cfg.auth = {
-        type: 'OAuth2',
-        oauth2: {
-          clientId: 'e6b02a7d-eb7e-4090-b112-f78f68cd6022',
-          clientSecret: 'e6b02a7d-eb7e-4090-b112-f78f68cd6022',
-          authUri: 'http://example.com/oauth/auth',
-          tokenUri: 'http://example.com/oauth/token',
-          keys: {
-            access_token: 'token',
-            token_type: 'Bearer',
-            refresh_token: 'refresh_token',
-            expires_in: 28800,
-            tokenExpiryTime: new Date().setDate(new Date().getDate() + 1),
-          },
-        },
-      };
-
-      const requestNock = nock(msg.body.url, {
-        reqheaders: {
-          Authorization: `Bearer ${cfg.auth.oauth2.keys.access_token}`,
-        },
-      })
-        .intercept('/', 'POST')
-        .reply((uri, requestBody) => [
-          200,
-          { success: true },
-        ]);
-
-      await processAction.call(emitter, msg, cfg);
-
-      expect(requestNock.isDone());
-    });
-
-    it('should refresh token with non-expiring refresh_token', async () => {
-      const refreshedToken = 'refreshed_token';
-      const tokenUri = 'http://example.com/oauth/token/';
-
-      cfg.auth = {
-        type: 'OAuth2',
-        oauth2: {
-          clientId: 'e6b02a7d-eb7e-4090-b112-f78f68cd6022',
-          clientSecret: 'e6b02a7d-eb7e-4090-b112-f78f68cd6022',
-          authUri: 'http://example.com/oauth/auth',
-          tokenUri,
-          keys: {
-            access_token: 'token',
-            token_type: 'Bearer',
-            refresh_token: 'refresh_token',
-          },
-        },
-      };
-
-      const responseMessage = {
-        access_token: refreshedToken,
-        token_type: 'Bearer',
-      };
-
-      const refreshTokenNock = nock(tokenUri)
-        .post('/', {
-          refresh_token: cfg.auth.oauth2.keys.refresh_token,
-          grant_type: 'refresh_token',
-          client_id: cfg.auth.oauth2.clientId,
-          client_secret: cfg.auth.oauth2.clientSecret,
-        })
-        .reply((uri, requestBody) => [
-          200,
-          responseMessage,
-        ]);
-
-
-      const requestNock = nock(msg.body.url, {
-        reqheaders: {
-          Authorization: `Bearer ${refreshedToken}`,
-        },
-      })
-        .intercept('/', 'POST')
-        .reply((uri, requestBody) => [
-          200,
-          { success: true },
-        ]);
-
-      await processAction.call(emitter, msg, cfg);
-
-      expect(refreshTokenNock.isDone());
-      expect(requestNock.isDone());
-    });
+  afterEach(() => {
+    delete process.env.NODE_TLS_REJECT_UNAUTHORIZED;
   });
 
   describe('split result', () => {
@@ -291,7 +87,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method: 'POST',
         },
-        auth: {},
+        secretId: '1234',
       };
       const responseMessage = ['first', 'second', 'third'];
       nock(JsonataTransform.jsonataTransform(msg,
@@ -301,12 +97,13 @@ describe('httpRequest action', () => {
           200,
           responseMessage,
         ]);
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       // eslint-disable-next-line no-unused-expressions
       expect(messagesNewMessageWithBodyStub.calledThrice).to.be.true;
-      expect(messagesNewMessageWithBodyStub.args[0][0]).to.be.eql('first');
-      expect(messagesNewMessageWithBodyStub.args[1][0]).to.be.eql('second');
-      expect(messagesNewMessageWithBodyStub.args[2][0]).to.be.eql('third');
+      expect(messagesNewMessageWithBodyStub.args[0][0]).to.deep.include({ body: 'first', statusCode: 200 });
+      expect(messagesNewMessageWithBodyStub.args[1][0]).to.deep.include({ body: 'second', statusCode: 200 });
+      expect(messagesNewMessageWithBodyStub.args[2][0]).to.deep.include({ body: 'third', statusCode: 200 });
     });
     it('should emit array of item if splitResult=false', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
@@ -322,7 +119,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method: 'POST',
         },
-        auth: {},
+        secretId: 55555,
       };
       const responseMessage = ['first', 'second', 'third'];
       nock(JsonataTransform.jsonataTransform(msg, { expression: cfg.reader.url }, emitter))
@@ -332,10 +129,12 @@ describe('httpRequest action', () => {
           200,
           responseMessage,
         ]);
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       // eslint-disable-next-line no-unused-expressions
       expect(messagesNewMessageWithBodyStub.calledOnce).to.be.true;
-      expect(messagesNewMessageWithBodyStub.args[0][0]).to.be.eql({ result: responseMessage });
+      expect(messagesNewMessageWithBodyStub.args[0][0]).to.deep
+        .include({ body: responseMessage, statusCode: 200 });
     });
     it('splitResult=true should be ignored if item is not array', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
@@ -351,7 +150,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method: 'POST',
         },
-        auth: {},
+        secretId: Math.random(),
       };
       const responseMessage = { data: 'not array' };
       nock(JsonataTransform.jsonataTransform(msg, { expression: cfg.reader.url }, emitter))
@@ -361,13 +160,14 @@ describe('httpRequest action', () => {
           200,
           responseMessage,
         ]);
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       // eslint-disable-next-line no-unused-expressions
       expect(messagesNewMessageWithBodyStub.calledOnce).to.be.true;
-      expect(messagesNewMessageWithBodyStub.args[0][0]).to.be.eql(responseMessage);
+      expect(messagesNewMessageWithBodyStub.args[0][0]).to.deep
+        .include({ body: responseMessage, statusCode: 200 });
     });
   });
-
 
   describe('when all params is correct', () => {
     ['GET', 'POST', 'PUT', 'PATCH', 'DELETE'].forEach((method, index) => {
@@ -385,7 +185,7 @@ describe('httpRequest action', () => {
             url: 'url',
             method,
           },
-          auth: {},
+          secretId: Math.random(),
         };
 
         const responseMessage = { message: `hello world ${index}` };
@@ -398,9 +198,10 @@ describe('httpRequest action', () => {
             responseMessage,
           ]);
 
+        setNoAuthSecretStub(cfg.secretId);
         await processAction.call(emitter, msg, cfg);
         expect(messagesNewMessageWithBodyStub.args[0][0])
-          .to.eql(responseMessage);
+          .to.deep.include({ body: responseMessage, statusCode: 200 });
       });
     });
     ['GET', 'POST', 'PUT', 'PATCH', 'DELETE'].forEach((method) => {
@@ -417,7 +218,7 @@ describe('httpRequest action', () => {
               },
             ],
           },
-          auth: {},
+          secretId: Math.random(),
         };
 
         if (method !== 'GET') {
@@ -448,7 +249,7 @@ describe('httpRequest action', () => {
               '{}',
             ];
           });
-
+        setNoAuthSecretStub(cfg.secretId);
         await processAction.call(emitter, msg, cfg);
       });
     });
@@ -470,7 +271,7 @@ describe('httpRequest action', () => {
             },
           ],
         },
-        auth: {},
+        secretId: Math.random(),
       };
 
       const responseMessage = 'hello world';
@@ -489,7 +290,7 @@ describe('httpRequest action', () => {
             responseMessage,
           ];
         });
-
+      setNoAuthSecretStub(cfg.secretId);
       processAction.call(emitter, msg, cfg);
     });
     it('should pass multiple headers properly', (done) => {
@@ -518,7 +319,7 @@ describe('httpRequest action', () => {
             },
           ],
         },
-        auth: {},
+        secretId: Math.random(),
       };
 
       const responseMessage = 'hello world';
@@ -540,6 +341,7 @@ describe('httpRequest action', () => {
           ];
         });
 
+      setNoAuthSecretStub(cfg.secretId);
       processAction.call(emitter, msg, cfg);
     });
     describe('when request body is passed', () => {
@@ -563,7 +365,7 @@ describe('httpRequest action', () => {
               encoding: 'raw',
             },
           },
-          auth: {},
+          secretId: Math.random(),
         };
 
         const responseMessage = 'hello world';
@@ -578,7 +380,7 @@ describe('httpRequest action', () => {
               responseMessage,
             ];
           });
-
+        setNoAuthSecretStub(cfg.secretId);
         processAction.call(emitter, msg, cfg);
       });
       it('should properly pass formdata body', (done) => {
@@ -612,7 +414,7 @@ describe('httpRequest action', () => {
             },
             headers: [],
           },
-          auth: {},
+          secretId: '1234',
         };
 
         const responseMessage = 'hello world';
@@ -627,7 +429,7 @@ describe('httpRequest action', () => {
               responseMessage,
             ];
           });
-
+        setNoAuthSecretStub(cfg.secretId);
         processAction.call(emitter, msg, cfg);
       });
     });
@@ -647,13 +449,14 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       nock(JsonataTransform.jsonataTransform(msg, { expression: cfg.reader.url }, emitter))
         .intercept('/', method)
         .delay(20 + Math.random() * 200)
         .replyWithError('something awful happened');
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg).catch((e) => {
         expect(e.message).to.be.eql('Error: something awful happened');
       });
@@ -676,14 +479,14 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       nock(JsonataTransform.jsonataTransform(msg, { expression: cfg.reader.url }, emitter))
         .intercept('/', method)
         .delay(20 + Math.random() * 200)
         .replyWithError('something awful happened');
-
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg).catch((e) => {
         expect(e.message).to.be.eql('Error: something awful happened');
         expect(emitter.emit.withArgs('rebound').callCount).to.be.equal(1);
@@ -704,7 +507,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       nock(JsonataTransform.jsonataTransform(msg, { expression: cfg.reader.url }, emitter))
@@ -712,6 +515,7 @@ describe('httpRequest action', () => {
         .delay(20 + Math.random() * 200)
         .reply(408, 'Error');
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(emitter.emit.withArgs('rebound').callCount).to.be.equal(1);
       expect(emitter.emit.withArgs('rebound').args[0][1]).to.be.equal(
@@ -732,9 +536,10 @@ describe('httpRequest action', () => {
         reader: {
           url: 'url',
         },
-        auth: {},
+        secretId: '1234',
       };
 
+      setNoAuthSecretStub(cfg.secretId);
       try {
         await processAction.call(emitter, msg, cfg);
       } catch (err) {
@@ -752,9 +557,10 @@ describe('httpRequest action', () => {
         reader: {
           method: 'GET',
         },
-        auth: {},
+        secretId: '1234',
       };
 
+      setNoAuthSecretStub(cfg.secretId);
       try {
         await processAction.call(emitter, msg, cfg);
       } catch (err) {
@@ -773,9 +579,9 @@ describe('httpRequest action', () => {
           url: 'url',
           method: 'GETT',
         },
-        auth: {},
+        secretId: '1234',
       };
-
+      setNoAuthSecretStub(cfg.secretId);
       try {
         await processAction.call(emitter, msg, cfg);
       } catch (err) {
@@ -803,7 +609,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       const responseMessage = '';
@@ -813,10 +619,12 @@ describe('httpRequest action', () => {
         .delay(20 + Math.random() * 200)
         .reply(204, responseMessage);
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
 
       // eslint-disable-next-line no-unused-expressions
-      expect(messagesNewMessageWithBodyStub.args[0][0]).to.exist;
+      expect(messagesNewMessageWithBodyStub.args[0][0]).to.deep
+        .include({ statusCode: 204, body: undefined });
     });
     it('No response body && dontThrowErrorFlg false', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
@@ -834,7 +642,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       const responseMessage = '';
@@ -844,10 +652,15 @@ describe('httpRequest action', () => {
         .delay(20 + Math.random() * 200)
         .reply(204, responseMessage);
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
 
       expect(messagesNewMessageWithBodyStub.lastCall.args[0])
-        .to.deep.equal({});
+        .to.deep.include({
+          headers: {},
+          body: undefined,
+          statusCode: 204,
+        });
     });
     it('Valid XML Response && dontThrowErrorFlg true', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
@@ -865,7 +678,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       nock(JsonataTransform.jsonataTransform(msg, { expression: cfg.reader.url }, emitter))
@@ -875,6 +688,7 @@ describe('httpRequest action', () => {
           'Content-Type': 'application/xml',
         });
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
 
       expect(messagesNewMessageWithBodyStub.lastCall.args[0])
@@ -903,7 +717,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       nock(JsonataTransform.jsonataTransform(msg, { expression: cfg.reader.url }, emitter))
@@ -912,13 +726,15 @@ describe('httpRequest action', () => {
         .reply(200, '<xml>foo</xml>', {
           'Content-Type': 'application/xml',
         });
-
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
 
       expect(messagesNewMessageWithBodyStub.lastCall.args[0])
-        .to.deep.equal(
-          { xml: 'foo' },
-        );
+        .to.deep.include({
+          body: { xml: 'foo' },
+          headers: { 'content-type': 'application/xml' },
+          statusCode: 200,
+        });
     });
     it('Invalid XML Response', async () => {
       const method = 'POST';
@@ -933,7 +749,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       nock(JsonataTransform.jsonataTransform(msg, { expression: cfg.reader.url }, emitter))
@@ -943,6 +759,7 @@ describe('httpRequest action', () => {
           'Content-Type': 'application/xml',
         });
 
+      setNoAuthSecretStub(cfg.secretId);
       try {
         await processAction.call(emitter, msg, cfg);
         throw new Error('This line should never be called because await above should throw an error');
@@ -966,7 +783,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       const responseMessage = 'boom!';
@@ -978,7 +795,7 @@ describe('httpRequest action', () => {
           200,
           responseMessage,
         ]);
-
+      setNoAuthSecretStub(cfg.secretId);
       try {
         await processAction.call(emitter, msg, cfg);
         throw new Error('This line should never be called because await above should throw an error');
@@ -1002,7 +819,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       const responseMessage = '{"id":"1", "name":"John", "surname":"Malkovich"}';
@@ -1014,6 +831,7 @@ describe('httpRequest action', () => {
           200,
           responseMessage,
         ]);
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to.deep.eql({
         headers: {},
@@ -1043,6 +861,7 @@ describe('httpRequest action', () => {
           method,
         },
         auth: {},
+        secretId: '1234',
       };
 
       const responseMessage = '{"id":"1", "name":"John", "surname":"Malkovich"}';
@@ -1054,14 +873,19 @@ describe('httpRequest action', () => {
           200,
           responseMessage,
         ]);
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
-      expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to.deep.eql(
+      expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to.deep.eql({
+        body:
         {
           id: '1',
           name: 'John',
           surname: 'Malkovich',
         },
-      );
+        statusCode: 200,
+        headers: {},
+        statusMessage: null,
+      });
     });
     it('XML string without content-type   && dontThrowErrorFlg false', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
@@ -1079,7 +903,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       const responseMessage = '<first>1</first><second>2</second>';
@@ -1091,8 +915,14 @@ describe('httpRequest action', () => {
           200,
           responseMessage,
         ]);
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
-      expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to.eql({ result: responseMessage });
+      expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to.eql({
+        body: responseMessage,
+        statusCode: 200,
+        statusMessage: null,
+        headers: {},
+      });
     });
     it('XML string without content-type   && dontThrowErrorFlg true', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
@@ -1110,7 +940,7 @@ describe('httpRequest action', () => {
           url: 'url',
           method,
         },
-        auth: {},
+        secretId: '1234',
       };
 
       const responseMessage = '<first>1</first><second>2</second>';
@@ -1122,9 +952,10 @@ describe('httpRequest action', () => {
           200,
           responseMessage,
         ]);
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to.deep.equal({
-        body: { result: responseMessage },
+        body: responseMessage,
         headers: {},
         statusCode: 200,
         statusMessage: null,
@@ -1150,7 +981,7 @@ describe('httpRequest action', () => {
         },
         followRedirect: 'followRedirects',
         dontThrowErrorFlg: true,
-        auth: {},
+        secretId: '1234',
       };
 
       nock('http://example.com')
@@ -1161,6 +992,7 @@ describe('httpRequest action', () => {
         .get('/Login')
         .reply(200, '{"state": "after redirection"}', { 'Content-Type': 'application/json' });
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to.deep.equal({
         body: {
@@ -1189,7 +1021,7 @@ describe('httpRequest action', () => {
           method,
         },
         followRedirect: 'followRedirects',
-        auth: {},
+        secretId: '1234',
       };
 
       nock('http://example.com')
@@ -1200,11 +1032,16 @@ describe('httpRequest action', () => {
         .get('/Login')
         .reply(200, '{"state": "after redirection"}', { 'Content-Type': 'application/json' });
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(messagesNewMessageWithBodyStub.lastCall.args[0])
         .to
         .deep
-        .equal({ state: 'after redirection' });
+        .include({
+          body: { state: 'after redirection' },
+          statusCode: 200,
+          headers: { 'content-type': 'application/json' },
+        });
     });
     it('redirect request false && dontThrowErrorFlg true', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
@@ -1223,7 +1060,7 @@ describe('httpRequest action', () => {
         },
         dontThrowErrorFlg: true,
         followRedirect: 'doNotFollowRedirects',
-        auth: {},
+        secretId: '1234',
       };
 
       nock('http://example.com')
@@ -1235,6 +1072,7 @@ describe('httpRequest action', () => {
         .get('/Login')
         .reply(200, '{"state": "after redirection"}', { 'Content-Type': 'application/json' });
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to.deep.equal({
         headers:
@@ -1263,7 +1101,7 @@ describe('httpRequest action', () => {
           method,
         },
         followRedirect: 'doNotFollowRedirects',
-        auth: {},
+        secretId: '1234',
       };
 
       nock('http://example.com')
@@ -1275,11 +1113,16 @@ describe('httpRequest action', () => {
         .get('/Login')
         .reply(200, '{"state": "after redirection"}', { 'Content-Type': 'application/json' });
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(messagesNewMessageWithBodyStub.lastCall.args[0])
         .to
         .deep
-        .equal({ state: 'before redirection' });
+        .include({
+          body: { state: 'before redirection' },
+          statusCode: 302,
+          headers: { location: 'http://example.com/Login', 'content-type': 'application/json' },
+        });
     });
     it('redirect request false POST && dontThrowErrorFlg false', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
@@ -1297,7 +1140,7 @@ describe('httpRequest action', () => {
           method,
         },
         followRedirect: 'doNotFollowRedirects',
-        auth: {},
+        secretId: '1234',
       };
 
       nock('http://example.com')
@@ -1307,13 +1150,18 @@ describe('httpRequest action', () => {
           'Content-Type': 'application/json',
         })
         .get('/Login')
-        .reply(200, '{"state": "after redirection"}', { 'Content-Type': 'application/json' });
+        .reply(200, '{"state": "after redirection"}', { 'content-type': 'application/json' });
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(messagesNewMessageWithBodyStub.lastCall.args[0])
         .to
         .deep
-        .equal({ state: 'before redirection' });
+        .include({
+          body: { state: 'before redirection' },
+          statusCode: 302,
+          headers: { location: 'http://example.com/Login', 'content-type': 'application/json' },
+        });
     });
     it('redirect request false POST && dontThrowErrorFlg false', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
@@ -1331,7 +1179,7 @@ describe('httpRequest action', () => {
           method,
         },
         followRedirect: 'followRedirects',
-        auth: {},
+        secretId: '1234',
       };
 
       nock('http://example.com')
@@ -1343,15 +1191,20 @@ describe('httpRequest action', () => {
         .get('/Login')
         .reply(200, '{"state": "after redirection"}', { 'Content-Type': 'application/json' });
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(messagesNewMessageWithBodyStub.lastCall.args[0])
         .to
         .deep
-        .equal({ state: 'after redirection' });
+        .include({
+          body: { state: 'after redirection' },
+          statusCode: 200,
+          headers: { 'content-type': 'application/json' },
+        });
     });
   });
   describe('attachments', () => {
-    it('action message with attachments', async () => {
+    it('action message with inbound attachments', async () => {
       const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
         .returns(Promise.resolve());
       const inputMsg = {
@@ -1407,7 +1260,7 @@ describe('httpRequest action', () => {
           },
           headers: [],
         },
-        auth: {},
+        secretId: '1234',
       };
 
       nock('http://qwre.com')
@@ -1420,11 +1273,68 @@ describe('httpRequest action', () => {
           200,
           rawString,
         ]);
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, inputMsg, cfg);
-      expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to.eql({ result: rawString });
+      expect(messagesNewMessageWithBodyStub.lastCall.args[0]).to
+        .include({ body: rawString, statusCode: 200 });
+    });
+
+    it('action message with outbound attachment', async () => {
+      const inputMsg = {
+        body: {},
+      };
+
+      const fileContents = fs.readFileSync('./logo.png');
+
+      const cfg = {
+        reader: {
+          url: '"https://example.com/image.png"',
+          method: 'GET',
+          headers: [],
+        },
+        auth: {},
+      };
+
+      sinon.stub(client.resources.storage, 'createSignedUrl').returns(Promise.resolve({
+        get_url: 'http://example.com/getUrl',
+        put_url: 'http://example.com/putUrl',
+      }));
+
+      nock('http://example.com')
+        .put('/putUrl')
+        .reply((uri, requestBody) => {
+          expect(requestBody).to.deep.equal(fileContents.toString('hex'));
+          return [
+            200,
+          ];
+        });
+
+      nock('https://example.com')
+        .get('/image.png')
+        .reply((uri, requestBody) => [
+          200,
+          fileContents,
+          {
+            'Content-Type': 'image/png',
+            'content-length': fileContents.length,
+          },
+        ]);
+
+      setNoAuthSecretStub(cfg.secretId);
+      await processAction.call(emitter, inputMsg, cfg);
+      expect(emitter.emit.args[0][1].body).to.deep.eql({
+        statusCode: 200,
+        statusMessage: null,
+        headers: { 'content-type': 'image/png', 'content-length': '22421' },
+        attachments: {
+          'content-type': 'image/png',
+          size: '22421',
+          sourceUrl: 'https://example.com/image.png',
+          url: 'http://example.com/getUrl',
+        },
+      });
     });
   });
-
 
   describe('404 not found', () => {
     it('404 not found && dontThrowErrorFlg true', async () => {
@@ -1448,9 +1358,10 @@ describe('httpRequest action', () => {
         },
         followRedirect: 'followRedirects',
         dontThrowErrorFlg: true,
-        auth: {},
+        secretId: '1234',
       };
 
+      setNoAuthSecretStub(cfg.secretId);
       await processAction.call(emitter, msg, cfg);
       expect(messagesNewMessageWithBodyStub.lastCall.args[0].statusCode).to.eql(404);
       // TODO: should be 'Not Found' but nock doesn't allow statusMessage to be mocked https://github.com/nock/nock/issues/469
@@ -1476,15 +1387,80 @@ describe('httpRequest action', () => {
         },
         followRedirect: 'followRedirects',
         dontThrowErrorFlg: false,
-        auth: {},
+        secretId: '1234',
       };
 
-      await processAction.call(emitter, msg, cfg).then((result) => {
-        throw new Error('Test case does not expect success response');
-      }).catch((e) => {
-        // TODO: should be 'Code: 404 Message: Not Found' but nock doesn't allow statusMessage to be mocked https://github.com/nock/nock/issues/469
-        expect(e.message).to.be.eql('Code: 404 Message: HTTP error');
-      });
+      setNoAuthSecretStub(cfg.secretId);
+      await processAction.call(emitter, msg, cfg);
+      expect(emitter.emit.callCount).to.equal(2);
+      expect(emitter.emit.args[0][0]).to.equal('error');
+      expect(emitter.emit.args[1][0]).to.equal('end');
+    });
+  });
+
+  describe('delay between calls', () => {
+    it('should wait delayBetweenCalls', async () => {
+      const messagesNewMessageWithBodyStub = stub(messages, 'newMessageWithBody')
+        .returns(Promise.resolve());
+      const msg = {
+        body: {
+          url: 'http://example.com',
+        },
+        passthrough: { test: 'test' },
+      };
+      const cfg = {
+        splitResult: true,
+        reader: {
+          url: 'url',
+          method: 'POST',
+        },
+        secretId: '1234',
+        delay: '20',
+        callCount: '4',
+      };
+      const responseMessage = ['first', 'second', 'third'];
+      nock(JsonataTransform.jsonataTransform(msg,
+        { expression: cfg.reader.url }, emitter))
+        .intercept('/', 'POST')
+        .reply((uri, requestBody) => [
+          200,
+          responseMessage,
+        ]);
+      setNoAuthSecretStub(cfg.secretId);
+      await processAction.call(emitter, msg, cfg);
+      // eslint-disable-next-line no-unused-expressions
+      expect(messagesNewMessageWithBodyStub.calledThrice).to.be.true;
+      expect(messagesNewMessageWithBodyStub.args[0][0].body).to.be.eql('first');
+      expect(messagesNewMessageWithBodyStub.args[1][0].body).to.be.eql('second');
+      expect(messagesNewMessageWithBodyStub.args[2][0].body).to.be.eql('third');
+    });
+  });
+
+  describe('timeout configuration', () => {
+    // Temporarily skip test as https://httpstat.us/200 is temporarily down
+    it.skip('should fail on small timeout', async () => {
+      const msg = {
+        body: {
+          url: 'https://httpstat.us/200?sleep=5000',
+        },
+        passthrough: { test: 'test' },
+      };
+      const cfg = {
+        splitResult: true,
+        reader: {
+          url: 'url',
+          method: 'GET',
+        },
+        auth: {},
+        requestTimeoutPeriod: '1000',
+      };
+
+      // Workaround for https://github.com/Readify/httpstatus/issues/79
+      process.env.NODE_TLS_REJECT_UNAUTHORIZED = 0;
+
+      await processAction.call(emitter, msg, cfg);
+      expect(emitter.emit.getCall(0).args[0]).to.be.equals('error');
+      expect(emitter.emit.getCall(0).args[1].message).to.be.equals(`Timeout error! Waiting for response more than ${cfg.requestTimeoutPeriod} ms`);
     });
   });
 });